+++ tests/testthat/test-class.R
@@ -114,9 +114,9 @@ describe("new_object()", {
     expect_snapshot(new_object(), error = TRUE)
   })
 
-  it("errors if `.parent` doesn't inherit from the parent class (#409)", {
+  it("errors if `_parent` doesn't inherit from the parent class (#409)", {
     Bar <- new_class("Bar", package = NULL)
-    # `.parent` should be `Bar()`, not the class spec `Bar`
+    # `_parent` should be `Bar()`, not the class spec `Bar`
     Foo <- new_class(
       "Foo",
       parent = Bar,
@@ -147,7 +147,7 @@ describe("new_object()", {
     expect_no_error(Concrete(x = 1L))
   })
 
-  it("errors if `.parent` is supplied but class has no parent", {
+  it("errors if `_parent` is supplied but class has no parent", {
     NoParent <- new_class(
       "NoParent",
       package = NULL,
@@ -155,6 +155,17 @@ describe("new_object()", {
       constructor = function() new_object(42L)
     )
     expect_snapshot(NoParent(), error = TRUE)
+  })
+
+  it("can set a property named `.parent` (#423)", {
+    foo <- new_class(
+      "foo",
+      properties = list(.parent = class_double),
+      package = NULL,
+      constructor = function(.parent) new_object(S7_object(), .parent = .parent)
+    )
+    obj <- foo(.parent = 1)
+    expect_equal(obj@.parent, 1)
   })
 
   it("validates object", {